requirements.txt
pipecat-ai[local,openai,elevenlabs]
python-dotenv
streamlit==1.28.0
openai==1.3.0
python-dotenv==1.0.0
audio-recorder-streamlit==0.0.6
pyaudio==0.2.11
numpy>=1.21.0
sounddevice>=0.4.0
soundfile>=0.10.0